changed type

+++ src/veda/vedaprofiling_types.h
@@ -10,7 +10,7 @@ typedef struct {
    double acc_DtoH_bw;
    double acc_H_bw;
    long long total_time_nsec;
-   size_t overhead_nsec;
+   long long overhead_nsec;
 } vedaprofile_t;
 
 #endif
